+++ lib/waterdrop/errors.rb
@@ -60,4 +60,12 @@ module WaterDrop
       end
     end
   end
+
+  # Alias so we can have a nicer API to abort transactions
+  # This makes referencing easier
+  AbortTransaction = Errors::AbortTransaction
+
+  # Alias so we can have a nicer API to abort transactions
+  # This mimics Rails convention
+  Rollback = Errors::AbortTransaction
 end