+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # WaterDrop changelog
 
 ## 2.7.4 (Unreleased)
-- [Fix] Consumer consuming error: Local: Erroneous state (state) post break flow in transaction.
+- [Maintenance] Alias `WaterDrop::Errors::AbortTransaction` with `WaterDrop::Rollback` and `WaterDrop::AbortTransaction`.
 - [Maintenance] Lower the precision reporting to 100 microseconds in the logger listener.
+- [Fix] Consumer consuming error: Local: Erroneous state (state) post break flow in transaction.
 - [Change] Require 'karafka-core' `>= 2.4.3`
 
 ## 2.7.3 (2024-06-09)